mv Config to lib and make public
