+++ rust_book/minigrep/src/main.rs
@@ -1,7 +1,7 @@
 use std::env;
-use std::fs;
 use std::process;
-use std::error::Error;
+
+use minigrep::Config;
 
 /// cargo run -- searchstring poem.txt
 fn main() {
@@ -18,34 +18,12 @@ fn main() {
     println!("Searching for {}", config.query);
     println!("In file {}", config.file_path);
 
-    if let Err(e) = run(config){
+    if let Err(e) = minigrep::run(config){
         println!("Should have been able to read the file - Application error: {e}");
         process::exit(1);
     }
 }
 
-struct Config {
-    query: String,
-    file_path: String,
-}
-impl Config {
-    fn build(args: &[String]) -> Result<Config, &'static str> {
-        if args.len() < 3 {
-            return Err("not enough arguments");
-        }
-        let query = args[1].clone();
-        let file_path = args[2].clone();
 
-        Ok(Config { query, file_path })
-    }
-}
-
-fn run(config: Config) -> Result<(), Box<dyn Error>>{
-    let contents = fs::read_to_string(config.file_path)?;
-        //.expect("Should have been able to read the file");
-
-    println!("With text:\n{contents}");
-    Ok(())
-}
